+++ tests/models/test_gift_card.py
@@ -74,4 +74,4 @@ def test_gift_card_rejects_negative_initial_amount() -> None:
 
 def to_cents(dollars: float) -> int:
     """Dollars to integer cents (ADR-0002)."""
-    return int(round(dollars * 100))
+    return round(dollars * 100)
